Reduce messageBox coupling
